split tooltip hooks into their own files
